Added EuclideanRational and first pass at Natural.
EuclideanRational is like Rational but works for any type that has
a EuclideanRing instance. Due to its generality it's a bit more
limited, but it does have its own typeclass instances.

Natural is a somewhat weird bird: it's a cross between Haskell's
Nat type (slightly more advanced than Zero | Succ, it uses a binary
encoding) and java.math.BigInteger. Instead of being base-2, it's
base-2^32: each of its digits is a UInt. Unlike BigInteger it doesn't
use an array, but rather is an immutable cons list of values.

Natural isn't really done yet... division hasn't been implemented
in the most general case and there are other oversights. The most
pressing problem it probably has is that since determining the
number of "digits" in a Natural is an O(n) operation, many
comparisons are slower than they need to be, and there are
optimization opportunities we miss out on.