Updates to spire.math.Natural.
This commit adds a bunch of new operators:

    comparisons: < <= > >=
    shifts: << >>
    mod and quotmod for UInt: % /%
	logical operators: | & ^

There is still no support for divisors beyond those that fit in a
single UInt. Division may prove to be the place where the cons list
structure of Natural is the worst fit, and we have to fall back to
using a temporary array.