Fixed Natural bugs and implemented benchmark.
When summing numbers, if your numbers are 128 bits or less, Natural
is faster than BigInt (tested for both 1K and 1M numbers). Beyond
that BigInt wins... it seems like BigInt has a lot of fixed overhead
but a much better curve as bit-count increases (APFloat is an even
more extreme example of this).

So, it seems like there's a niche for Natural, although it may not
be a gigantic one. For posterity here are the rough ratios:

      8 bits: 4.1x faster
	 16 bits: 3.9x faster
	 32 bits: 2.1x faster
	 64 bits: 1.7x faster
	 96 bits: 1.3x faster
	128 bits: 1.1x faster
	196 bits: 1.4x SLOWER
	256 bits: 1.9x SLOWER

Obviously there's a lot of room for optimization. Some kind of hybrid
strategy that combines wrapped Longs (for <=64 bits), cons lists
(for <=128 bits), and arrays (for everything else) might end up being
the best.