Backport the current Spire code to 2.9.1.
This is done primarily to assist algebird integration, although it will
also be handy for anyone else wanting to use Spire on 2.9.

At this moment, all the tests are passing. However, since 2.9 contained
some bugs (particularly around specialization) compared to 2.10, I need
to make sure there aren't any serious performance regressions.

The biggest downside to the 2.9 version is that using infix operators
will incur a slight penalty compared to manually wrangling type class
instances. This is because we can't use macros (like in 2.10) to rewrite
the AST. Also, we lose the benefits of value classes, which impacts the
unsigned types as well as Natural. Finally, we lose the special syntax
support (which relies on macros and interpolation).

This version will result in 0.2.x (probably 0.2.5) since 0.3.x is
reserved for 2.10.

TODO:
1. Re-specialize Ops classes
2. Benchmarking
3. Update documentation
4. Update version
5. Release to Sonatype