Another backport back to 2.9.
This commit finalizes the backport of changes from the 2.10.0 branch.
It includes all work on that branch through c0c498b3d.

In particular this commit fixes lots of merge problems with the
previous N cherry-picks ignored. At this point everything is passing
except for some of the scalacheck law stuff which I am hoping Lars
can help fix.