Fix bugs in random forest example + yeast dataset.
There were several bugs in the random forest example that were
discovered while testing quality of the random forest on 2 data sets.
This also adds a new data set to cross-validate as an example.
Additionally, the data sets are moved out into their own file.

On top of others, the decision to always split a region with more than
minSplitSize points was made. Before, if there error couldn't be
decreased, then we gave up. However, since axes are sampled, this was a
bad strategy. Especially for datasets like the S. cerevisiae one, where
a couple of axes are 0.0 for a majority of data points.